requirements.txt
Django==1.4.1
Fabric==1.4.3
Pillow==1.7.7
South==0.7.6
amqplib==1.0.2
anyjson==0.3.3
beautifulsoup4==4.1.2
billiard==2.7.3.12
celery==3.0.6
coverage==3.5.2
django-appconf==0.5
django-celery==3.0.6
django-compressor==1.2
django-debug-toolbar==0.9.4
django-social-auth==0.7.4
facebook-sdk==0.3.2
gunicorn==0.14.6
httplib2==0.7.4
ipython==0.13
kombu==2.4.0
logilab-astng==0.24.0
logilab-common==0.58.1
numpy==1.6.2
oauth==1.0.1
oauth2==1.5.211
pycrypto==2.6
pylint==0.25.2
python-dateutil==1.5
python-memcached==1.48
python-openid==2.2.5
python-twitpic==2.1
requests==0.13.6
simplejson==2.6.1
ssh==1.7.14
tweepy==1.11
wsgiref==0.1.2
